Upgrade to tailwind 1.0 and get make build working again. This commit does not contain all our custom configuration yet.
