+++ resources/scss/components/_formy.scss
@@ -166,7 +166,7 @@
     }
 
     .login-info ul {
-        @apply .m-0 .p-0 .list-reset;
+        @apply .m-0 .p-0;
     }
 
     .formy_logout a,
