+++ resources/scss/components/_menu-main.scss
@@ -1,6 +1,6 @@
 .main-menu {
     ul {
-        @apply .list-reset .m-0 .mb-2 .p-0;
+        @apply .m-0 .mb-2 .p-0;
     }
 
     // Padding above and below menu
