+++ resources/scss/components/_menu-offcanvas.scss
@@ -1,5 +1,5 @@
 .slideout-menu {
-    @apply .bg-green-dark .fixed .pin-t .pin-r .pin-b;
+    @apply .bg-green-dark .fixed .top-0 .right-0 .bottom-0;
 
     width: 250px;
     height: 100%;
