+++ resources/scss/components/_responsive-embed.scss
@@ -8,7 +8,7 @@
 .responsive-embed object,
 .responsive-embed embed,
 .responsive-embed video {
-    @apply .absolute .pin-t .pin-l .w-full .h-full;
+    @apply .absolute .top-0 .left-0 .w-full .h-full;
 }
 
 .responsive-embed.widescreen {
