+++ resources/scss/components/_skip.scss
@@ -7,5 +7,5 @@
 .skip a:active,
 .skip a:focus,
 .skip a:hover {
-    @apply .pin-l .pin-t .overflow-visible .z-50 .w-auto .h-auto;
+    @apply .left-0 .top-0 .overflow-visible .z-50 .w-auto .h-auto;
 }
