+++ resources/scss/components/_text-shadow.scss
@@ -1,5 +1,5 @@
 @responsive {
     .text-shadow-dark {
-        text-shadow: 0 1px 0 config('colors.green-darkest');
+        text-shadow: 0 1px 0 theme('colors.green-darkest');
     }
 }
