+++ resources/views/components/breadcrumbs.blade.php
@@ -2,7 +2,7 @@
     $breadcrumbs => array // ['display_name', 'relative_url']
 --}}
 <nav class="mt-6 mb-2" aria-label="Breadcrumbs">
-    <ul class="list-reset text-sm">
+    <ul class="text-sm">
         @foreach($breadcrumbs as $key=>$crumb)
             @if($key == 0)
                 <li class="inline">
