+++ resources/views/components/button-svg-overlay-dark.blade.php
@@ -4,8 +4,8 @@
 
 @if(!empty($button['link']) && !empty($button['relative_url']) && !empty($button['secondary_relative_url']))
     <a href="{{ $button['link'] }}" class="block min-w-full relative rounded bg-cover bg-green-darkest" style="padding-top: 36.39%; background-image: url('{{ $button['relative_url'] }}'); ">
-        <div class="absolute pin p-4 rounded bg-green-darkest opacity-65"></div>
-        <div class="absolute min-w-full pin rounded">
+        <div class="absolute inset-0 p-4 rounded bg-green-darkest opacity-65"></div>
+        <div class="absolute min-w-full inset-0 rounded">
             @image($button['secondary_relative_url'], $button['title'])
         </div>
     </a>
