+++ resources/views/components/skip.blade.php
@@ -1,5 +1,5 @@
 <nav aria-label="Skip navigation" class="skip">
-    <ul class="list-reset">
+    <ul>
         @if(config('base.top_menu_enabled') === true)
             <li class="skip-site-menu"><a href="#top-menu">Skip to site menu</a></li>
         @endif
